backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
torch
scikit-learn
xgboost
joblib
langchain
langsmith
langchain-groq
requests
python-dotenv 
pandas
lightgbm
tenacity
langgraph
six
kafka-python